@@ -1,34 +1,44 @@
 name: budinoh
 description: Automatic build (apk/appbundle/ipa) and upload your project build to Firebase, PlayStore and AppStore.
-version: 0.0.6
+version: 1.0.0
 repository: https://github.com/BreX900/budinoh
 issue_tracker: https://github.com/BreX900/budinoh/issues
 
 # dart pub publish --dry-run
 
 environment:
-  sdk: '>=2.14.0 <3.0.0'
+  sdk: '>=2.19.0 <4.0.0'
+
+scripts:
+  # To generate *.g.dart files
+  build:runner: dart run build_runner watch --delete-conflicting-outputs
 
 dependencies:
-  args: ^2.3.0
+  args: ^2.4.1
   console: ^4.1.0
-  collection: ^1.15.0
+  collection: ^1.17.0
 
-  yaml: ^3.1.0
-  checked_yaml: ^2.0.1
-  json_annotation: ^4.1.0
+  yaml: ^3.1.2
+  checked_yaml: ^2.0.3
+  json_annotation: ^4.8.1
 
   path: ^1.8.0
 
-  pubspec: ^2.0.1
-  process_run: ^0.12.2+1
-  queue: ^3.1.0
+  pubspec: ^2.3.0
+  process_run: ^0.13.0
+  queue: ^3.1.0+2
+
+  googleapis_auth: ^1.4.0
+  googleapis: ^11.0.0
+
+  http: ^0.13.6
+  _discoveryapis_commons: ^1.0.5
 
 dev_dependencies:
-  lints: ^1.0.1
+  lints: ^2.0.1
 
   # To generate .g files:
   # dart run build_runner build --delete-conflicting-outputs
   # dart run build_runner watch --delete-conflicting-outputs
-  build_runner: ^2.1.2
-  json_serializable: ^5.0.2
+  build_runner: ^2.3.3
+  json_serializable: ^6.6.2
