@@ -1,49 +1,70 @@
+import 'package:budinoh/src/settings_serializable.dart';
 import 'package:budinoh/src/stringify.dart';
-import 'package:json_annotation/json_annotation.dart';
 
 part 'distribution_settings.g.dart';
 
-@JsonSerializable()
-class FirebaseSettings with Stringify {
+@SettingsSerializable()
+class FirebaseCliSettings with Stringify {
   final String appId;
   final List<String> groups;
 
-  const FirebaseSettings({
+  const FirebaseCliSettings({
     required this.appId,
     this.groups = const <String>[],
   });
 
-  factory FirebaseSettings.fromJson(Map<dynamic, dynamic> map) => _$FirebaseSettingsFromJson(map);
+  factory FirebaseCliSettings.fromJson(Map<dynamic, dynamic> map) =>
+      _$FirebaseCliSettingsFromJson(map);
   @override
-  Map<String, dynamic> toJson() => _$FirebaseSettingsToJson(this);
+  Map<String, dynamic> toJson() => _$FirebaseCliSettingsToJson(this);
 }
 
-// xcrun altool --upload-app -f <PathToFile>.ipa -u <Username> -p <Password>
-@JsonSerializable()
-class AppleStoreSettings with Stringify {
-  final String username;
-  final String password;
+@SettingsSerializable()
+class FirebaseApiSettings with Stringify {
+  final String? credentials;
+  final String appId;
+  final List<String> groups;
 
-  const AppleStoreSettings({
-    required this.username,
-    required this.password,
+  const FirebaseApiSettings({
+    this.credentials,
+    required this.appId,
+    this.groups = const <String>[],
   });
 
-  factory AppleStoreSettings.fromJson(Map<dynamic, dynamic> map) =>
-      _$AppleStoreSettingsFromJson(map);
+  factory FirebaseApiSettings.fromJson(Map<dynamic, dynamic> map) =>
+      _$FirebaseApiSettingsFromJson(map);
   @override
-  Map<String, dynamic> toJson() => _$AppleStoreSettingsToJson(this);
+  Map<String, dynamic> toJson() => _$FirebaseApiSettingsToJson(this);
 }
 
-// https://developers.google.com/android-publisher/api-ref/rest/v3/edits/insert
-// https://developers.google.com/android-publisher/api-ref/rest/v3/edits.bundles/upload
-// https://developers.google.com/android-publisher/api-ref/rest/v3/edits/commit
-@JsonSerializable()
+@SettingsSerializable()
+class AppleStoreAppSettings with Stringify {
+  final String apiIssuer;
+  final String apiKeyId;
+
+  /// Directory to key or key content
+  final String? apiKey;
+
+  const AppleStoreAppSettings({
+    required this.apiKeyId,
+    required this.apiIssuer,
+    this.apiKey,
+  });
+
+  factory AppleStoreAppSettings.fromJson(Map<dynamic, dynamic> map) =>
+      _$AppleStoreAppSettingsFromJson(map);
+  @override
+  Map<String, dynamic> toJson() => _$AppleStoreAppSettingsToJson(this);
+}
+
+@SettingsSerializable()
 class GoogleStoreSettings with Stringify {
-  final String apiKey;
+  final String? credentials;
+  final String packageName;
 
   const GoogleStoreSettings({
-    required this.apiKey,
+    this.credentials,
+    required this.packageName,
   });
 
   factory GoogleStoreSettings.fromJson(Map<dynamic, dynamic> map) =>
