@@ -1,61 +1,65 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:budinoh/src/d_e.dart';
-import 'package:path/path.dart';
+import 'package:budinoh/src/shell_extensions.dart';
 import 'package:process_run/shell.dart';
 
 class BuildClient {
   final shell = Shell();
 
+  final String? env;
+  final String? buildNumber;
+  final String? buildName;
+
+  BuildClient({
+    this.env,
+    this.buildNumber,
+    this.buildName,
+  });
+
   /// Call it every time it starts a build series with all different types
-  Future<void> initProject() async {
+  Future<void> cleanAndPubGet() async {
     await _flutterClean();
     await _flutterPubGet();
   }
 
-  /// Build apk
-  Future<File> buildApk(String envName, String env, Directory outputDir) async {
-    await shell.singleRun('flutter build apk --no-pub $env');
+  Future<File> buildApk({String? env, String? args}) async {
+    await shell.singleRun('flutter build apk --no-pub', _buildArgs(env: env, args: args));
 
-    // Move android build into output dir
-    final buildFile = File('build/app/outputs/flutter-apk/app-release.apk');
-    final buildFile1 = await buildFile.rename('${outputDir.path}/${basename(buildFile.path)}');
-    return File(buildFile1.path);
+    return File('build/app/outputs/flutter-apk/app-release.apk');
   }
 
-  /// Build appbundle
-  Future<File> buildAppBundle(String envName, String env, Directory outputDir) async {
-    await shell.singleRun('flutter build appbundle --no-pub $env');
+  Future<File> buildAppBundle({String? env, String? args}) async {
+    await shell.singleRun('flutter build appbundle --no-pub', _buildArgs(env: env, args: args));
 
-    // Move android build into output dir
-    final buildFile = File('build/app/outputs/bundle/release/app-release.aab');
-    return await buildFile.rename('${outputDir.path}/${basename(buildFile.path)}');
+    return File('build/app/outputs/bundle/release/app-release.aab');
   }
 
-  /// Build ipa
-  Future<File> buildIpa(
-    String envName,
-    String env,
-    Directory outputDir, {
-    required String exportOptions,
+  Future<File> buildIpa({
+    String? exportOptions,
+    String? env,
+    String? args,
   }) async {
-    await shell.singleRun('flutter build ipa --no-pub --export-options-plist=$exportOptions $env');
+    await shell.singleRun('flutter build ipa --no-pub', [
+      if (exportOptions != null) '--export-options-plist=$exportOptions',
+      ..._buildArgs(env: env, args: args),
+    ]);
 
-    // Move ios build into output dir
     final buildDir = await Directory('build/ios/ipa').list().toList();
-    final buildFile = buildDir.singleWhere((file) => file.path.endsWith('ipa'));
-    final buildFile1 = await buildFile.rename('${outputDir.path}/${basename(buildFile.path)}');
-    return File(buildFile1.path);
-  }
-
-  /// Call when all builds is completed
-  Future<void> disposeProject() async {
-    await _flutterClean();
-    await _flutterPubGet();
+    return File(buildDir.singleWhere((file) => file.path.endsWith('.ipa')).path);
   }
 
   Future<void> _flutterClean() async => await shell.singleRun('flutter clean');
 
   Future<void> _flutterPubGet() async => await shell.singleRun('flutter pub get');
+
+  Iterable<String> _buildArgs({String? env, String? args}) {
+    env ??= this.env;
+    return [
+      if (buildNumber != null) '--build-number=${buildNumber!}',
+      if (buildName != null) '--build-name=${buildName!}',
+      if (env != null) env,
+      if (args != null) args,
+    ];
+  }
 }
