@@ -1,6 +1,5 @@
 import 'dart:io';
 
-import 'package:budinoh/src/d_e.dart';
 import 'package:process_run/shell.dart';
 import 'package:pubspec/pubspec.dart';
 
@@ -11,39 +10,30 @@ class ProjectClient {
     String? directory,
   }) : _shell = Shell(workingDirectory: directory);
 
-  Future<String> readEnv({String file = '', String settings = ''}) async {
-    final command = StringBuffer('dart pub global run define_env');
-    if (file.isNotEmpty) {
-      command.write(' -f $file');
-    }
-    if (settings.isNotEmpty) {
-      command.write(' -s $settings');
-    }
-
-    final res = await _shell.singleRun(command.toString());
-
-    return res.outText;
+  Future<String> readDartDefineFile({required String filePath}) async {
+    if (!await File(filePath).exists()) throw StateError('Env file not exist $filePath');
+    return '--dart-define-from-file=$filePath';
   }
 
   Future<Map<String, String>> readReleaseNotes() async {
     final changeLog = await File('CHANGELOG.md').readAsString();
 
-    final _versions = <String, List<String>>{};
-    String _version = '';
+    final versions = <String, List<String>>{};
+    String version = '';
     for (var line in changeLog.split('\n')) {
       line = line.trim();
       if (line.isEmpty) continue;
 
       final versionMatch = RegExp(r'## (\d+\.\d+\.\d+)').firstMatch(line);
       if (versionMatch != null) {
-        _version = versionMatch.group(1)!;
-        _versions[_version] = [];
+        version = versionMatch.group(1)!;
+        versions[version] = [];
       }
 
-      _versions[_version]?.add(line);
+      versions[version]?.add(line);
     }
 
-    return _versions.map((key, value) => MapEntry(key, value.join('\n')));
+    return versions.map((key, value) => MapEntry(key, value.join('\n')));
   }
 
   Future<String> readVersion() async {
