@@ -1,20 +1,26 @@
-import 'package:budinoh/src/settings/build_settings.dart';
-import 'package:budinoh/src/stringify.dart';
+import 'package:budinoh/budinoh.dart';
+import 'package:budinoh/src/settings_serializable.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 part 'package_settings.g.dart';
 
-@JsonSerializable()
+@SettingsSerializable()
 class BuildGroupSettings with Stringify {
   final ApkSettings? apk;
   final IpaSettings? ipa;
   final AppBundleSettings? appbundle;
 
-  List<BuildSettings> get builds => [
-        if (apk != null) apk!,
-        if (ipa != null) ipa!,
-        if (appbundle != null) appbundle!,
-      ];
+  List<BuildSettings> getBuilds({
+    bool apk = true,
+    bool appbundle = true,
+    bool ipa = true,
+  }) {
+    return [
+      if (apk && this.apk != null) this.apk!,
+      if (appbundle && this.appbundle != null) this.appbundle!,
+      if (ipa && this.ipa != null) this.ipa!,
+    ];
+  }
 
   const BuildGroupSettings({
     this.apk,
@@ -29,13 +35,17 @@ class BuildGroupSettings with Stringify {
   Map<String, dynamic> toJson() => _$BuildGroupSettingsToJson(this);
 }
 
-@JsonSerializable()
+@SettingsSerializable()
 class PackageSettings with Stringify {
-  final String? defineEnv;
+  final EnvSettings? env;
+
+  final String? firebaseApiCredentials;
+
   final Map<String, BuildGroupSettings> builds;
 
   const PackageSettings({
-    this.defineEnv,
+    this.env,
+    this.firebaseApiCredentials,
     required this.builds,
   });
 
